Move timemachine buttons to add more room for graphs
