+++ ui/src/shared/components/TimeMachineQueries.tsx
@@ -1,10 +1,13 @@
 // Libraries
-import React, {SFC} from 'react'
+import React, {PureComponent} from 'react'
 import {connect} from 'react-redux'
 
 // Components
 import TimeMachineFluxEditor from 'src/shared/components/TimeMachineFluxEditor'
+import CSVExportButton from 'src/shared/components/CSVExportButton'
 import TimeMachineQueriesSwitcher from 'src/shared/components/TimeMachineQueriesSwitcher'
+import TimeMachineRefreshDropdown from 'src/shared/components/TimeMachineRefreshDropdown'
+import TimeRangeDropdown from 'src/shared/components/TimeRangeDropdown'
 import TimeMachineQueryTab from 'src/shared/components/TimeMachineQueryTab'
 import TimeMachineQueryBuilder from 'src/shared/components/TimeMachineQueryBuilder'
 import TimeMachineInfluxQLEditor from 'src/shared/components/TimeMachineInfluxQLEditor'
@@ -12,14 +15,21 @@ import TimeMachineQueriesTimer from 'src/shared/components/TimeMachineQueriesTim
 import SubmitQueryButton from 'src/shared/components/SubmitQueryButton'
 import {
   Button,
+  ComponentSpacer,
+  Alignment,
   ComponentColor,
   ComponentSize,
   ButtonShape,
   IconFont,
+  SlideToggle,
 } from 'src/clockface'
 
 // Actions
 import {addQuery} from 'src/shared/actions/v2/timeMachines'
+import {
+  setTimeRange,
+  setIsViewingRawData,
+} from 'src/shared/actions/v2/timeMachines'
 
 // Utils
 import {
@@ -36,6 +46,7 @@ import {
   DashboardQuery,
   InfluxLanguage,
   QueryEditMode,
+  TimeRange,
 } from 'src/types/v2'
 import {DashboardDraftQuery} from 'src/types/v2/dashboards'
 import {QueriesState} from 'src/shared/components/TimeSeries'
@@ -43,10 +54,14 @@ import {QueriesState} from 'src/shared/components/TimeSeries'
 interface StateProps {
   activeQuery: DashboardQuery
   draftQueries: DashboardDraftQuery[]
+  timeRange: TimeRange
+  isViewingRawData: boolean
 }
 
 interface DispatchProps {
   onAddQuery: typeof addQuery
+  onSetTimeRange: typeof setTimeRange
+  onSetIsViewingRawData: typeof setIsViewingRawData
 }
 
 interface OwnProps {
@@ -55,62 +70,98 @@ interface OwnProps {
 
 type Props = StateProps & DispatchProps & OwnProps
 
-const TimeMachineQueries: SFC<Props> = props => {
-  const {activeQuery, queriesState, draftQueries, onAddQuery} = props
+class TimeMachineQueries extends PureComponent<Props> {
+  public render() {
+    const {
+      queriesState,
+      draftQueries,
+      onAddQuery,
+      timeRange,
+      onSetTimeRange,
+      isViewingRawData,
+      queriesState: {files},
+    } = this.props
 
-  let queryEditor
-
-  if (activeQuery.editMode === QueryEditMode.Builder) {
-    queryEditor = <TimeMachineQueryBuilder />
-  } else if (activeQuery.type === InfluxLanguage.Flux) {
-    queryEditor = <TimeMachineFluxEditor />
-  } else if (activeQuery.type === InfluxLanguage.InfluxQL) {
-    queryEditor = <TimeMachineInfluxQLEditor />
+    return (
+      <div className="time-machine-queries">
+        <div className="time-machine-queries--controls">
+          <div className="time-machine-queries--tabs">
+            {draftQueries.map((query, queryIndex) => (
+              <TimeMachineQueryTab
+                key={queryIndex}
+                queryIndex={queryIndex}
+                query={query}
+              />
+            ))}
+            <Button
+              customClass="time-machine-queries--new"
+              shape={ButtonShape.Square}
+              icon={IconFont.PlusSkinny}
+              size={ComponentSize.ExtraSmall}
+              color={ComponentColor.Default}
+              onClick={onAddQuery}
+            />
+          </div>
+          <div className="time-machine-queries--buttons">
+            <TimeMachineQueriesTimer
+              status={queriesState.loading}
+              duration={queriesState.duration}
+            />
+            <ComponentSpacer align={Alignment.Right}>
+              <SlideToggle.Label text="View Raw Data" />
+              <SlideToggle
+                active={isViewingRawData}
+                onChange={this.handleToggleIsViewingRawData}
+                size={ComponentSize.ExtraSmall}
+              />
+              <TimeMachineRefreshDropdown />
+              <TimeRangeDropdown
+                timeRange={timeRange}
+                onSetTimeRange={onSetTimeRange}
+              />
+              <TimeMachineQueriesSwitcher />
+              <CSVExportButton files={files} />
+              <SubmitQueryButton queryStatus={queriesState.loading} />
+            </ComponentSpacer>
+          </div>
+        </div>
+        <div className="time-machine-queries--body">{this.queryEditor}</div>
+      </div>
+    )
   }
 
-  return (
-    <div className="time-machine-queries">
-      <div className="time-machine-queries--controls">
-        <div className="time-machine-queries--tabs">
-          {draftQueries.map((query, queryIndex) => (
-            <TimeMachineQueryTab
-              key={queryIndex}
-              queryIndex={queryIndex}
-              query={query}
-            />
-          ))}
-          <Button
-            customClass="time-machine-queries--new"
-            shape={ButtonShape.Square}
-            icon={IconFont.PlusSkinny}
-            size={ComponentSize.ExtraSmall}
-            color={ComponentColor.Default}
-            onClick={onAddQuery}
-          />
-        </div>
-        <div className="time-machine-queries--buttons">
-          <TimeMachineQueriesTimer
-            status={queriesState.loading}
-            duration={queriesState.duration}
-          />
-          <TimeMachineQueriesSwitcher />
-          <SubmitQueryButton queryStatus={queriesState.loading} />
-        </div>
-      </div>
-      <div className="time-machine-queries--body">{queryEditor}</div>
-    </div>
-  )
+  private get queryEditor(): JSX.Element {
+    const {activeQuery} = this.props
+
+    if (activeQuery.editMode === QueryEditMode.Builder) {
+      return <TimeMachineQueryBuilder />
+    } else if (activeQuery.type === InfluxLanguage.Flux) {
+      return <TimeMachineFluxEditor />
+    } else if (activeQuery.type === InfluxLanguage.InfluxQL) {
+      return <TimeMachineInfluxQLEditor />
+    }
+  }
+
+  private handleToggleIsViewingRawData = () => {
+    const {isViewingRawData, onSetIsViewingRawData} = this.props
+
+    onSetIsViewingRawData(!isViewingRawData)
+  }
 }
 
 const mstp = (state: AppState) => {
-  const {draftQueries} = getActiveTimeMachine(state)
+  const {draftQueries, timeRange, isViewingRawData} = getActiveTimeMachine(
+    state
+  )
   const activeQuery = getActiveQuery(state)
 
-  return {activeQuery, draftQueries}
+  return {timeRange, activeQuery, draftQueries, isViewingRawData}
 }
 
 const mdtp = {
   onAddQuery: addQuery,
+  onSetTimeRange: setTimeRange,
+  onSetIsViewingRawData: setIsViewingRawData,
 }
 
 export default connect<StateProps, DispatchProps, OwnProps>(
